logging-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
pika==1.3.2
flasgger==0.9.7.1